daily/20191112/example_graphql/requirements.txt
strawberry-graphql
ariadne
